refactor(api): sets → src/api modules (Effort 1)
Mechanical move of the sets feature from src/hooks/queries/sets to src/api/sets.
Added queryOptions factories for the set query hooks. No behavior change.
