+++ src/api/sets/useSetBySlug.ts
@@ -1,6 +1,6 @@
-import { useQuery } from "@tanstack/react-query";
+import { queryOptions, useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
-import { FestivalSet, setsKeys } from "./useSets";
+import { FestivalSet, setsKeys } from "./types";
 
 async function fetchSetBySlug({
   slug,
@@ -51,14 +51,19 @@ async function fetchSetBySlug({
   return transformedData;
 }
 
-// Hook
+export function setBySlugQuery(slug: string, editionId: string) {
+  return queryOptions({
+    queryKey: setsKeys.bySlug({ slug, editionId }),
+    queryFn: () => fetchSetBySlug({ slug, editionId }),
+  });
+}
+
 export function useSetBySlugQuery({
   slug,
   editionId,
 }: { slug?: string; editionId?: string } = {}) {
   return useQuery({
-    queryKey: setsKeys.bySlug({ slug, editionId }),
-    queryFn: () => fetchSetBySlug({ slug: slug!, editionId: editionId! }),
+    ...setBySlugQuery(slug!, editionId!),
     enabled: !!slug && !!editionId,
   });
 }
