+++ src/hooks/useScheduleData.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { formatDateTime } from "@/lib/timeUtils";
 import { format, startOfDay } from "date-fns";
-import type { FestivalSet } from "@/hooks/queries/sets/useSets";
+import type { FestivalSet } from "@/api/sets/types";
 import { Stage } from "./queries/stages/types";
 import { sortStagesByOrder } from "@/lib/stageUtils";
