+++ src/pages/SetDetails/SetInfoCard.tsx
@@ -9,7 +9,7 @@ import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import { Clock, ExternalLink, Music, Play } from "lucide-react";
 import { SetVotingButtons } from "./SetVotingButtons";
-import { FestivalSet } from "@/hooks/queries/sets/useSets";
+import { FestivalSet } from "@/api/sets/types";
 import { formatDayOnly, formatTimeRange } from "@/lib/timeUtils";
 import { GenreBadge } from "@/components/GenreBadge";
 import { StagePin } from "@/components/StagePin";
