+++ src/pages/admin/festivals/SetsManagement/SetManagement.tsx
@@ -3,9 +3,9 @@ import { useParams } from "@tanstack/react-router";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Loader2, Plus, Music } from "lucide-react";
-import { FestivalSet } from "@/hooks/queries/sets/useSets";
-import { useSetsByEditionQuery } from "@/hooks/queries/sets/useSetsByEdition";
-import { useDeleteSetMutation } from "@/hooks/queries/sets/useDeleteSet";
+import { FestivalSet } from "@/api/sets/types";
+import { useSetsByEditionQuery } from "@/api/sets/useSetsByEdition";
+import { useDeleteSetMutation } from "@/api/sets/useDeleteSet";
 import { useFestivalEditionBySlugQuery } from "@/api/editions/useFestivalEditionBySlug";
 import { SetFormDialog } from "../SetFormDialog";
 import { SetsTable } from "../SetsTable";
